Refactor Clan
