+++ rc-mobile-project/rc-mobile-project-dao/src/main/java/pe/com/rc/mobile/dao/ClanDAO.java
@@ -3,8 +3,8 @@ package pe.com.rc.mobile.dao;
 import java.util.List;
 
 import pe.com.rc.mobile.dao.helper.GenericDAO;
-import pe.com.rc.mobile.model.Clan;
 import pe.com.rc.mobile.model.User;
+import pe.com.rc.mobile.model.clan.Clan;
 
 public interface ClanDAO extends GenericDAO<Clan> {
