+++ rc-mobile-project/rc-mobile-project-model/src/main/java/pe/com/rc/mobile/model/Solicitude.java
@@ -9,6 +9,8 @@ import javax.persistence.JoinColumn;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
 
+import pe.com.rc.mobile.model.clan.Clan;
+
 @Entity
 @Table(name = "solicitude")
 public class Solicitude extends Record implements Serializable {
