+++ rc-mobile-project/rc-mobile-project-model/src/main/java/pe/com/rc/mobile/model/UserClanId.java
@@ -6,6 +6,8 @@ import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
 import javax.persistence.ManyToOne;
 
+import pe.com.rc.mobile.model.clan.Clan;
+
 @Embeddable
 public class UserClanId implements Serializable {
