+++ rc-mobile-project/rc-mobile-project-services/src/main/java/pe/com/rc/mobile/service/clan/ClanService.java
@@ -2,9 +2,9 @@ package pe.com.rc.mobile.service.clan;
 
 import java.util.List;
 
-import pe.com.rc.mobile.model.Clan;
 import pe.com.rc.mobile.model.User;
 import pe.com.rc.mobile.model.clan.BuildClanRequest;
+import pe.com.rc.mobile.model.clan.Clan;
 
 public interface ClanService {
