+++ rc-mobile-project/rc-mobile-project-web/src/main/java/pe/com/rc/mobile/web/controller/ClanController.java
@@ -9,9 +9,9 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import pe.com.rc.mobile.core.exception.ServiceException;
-import pe.com.rc.mobile.model.Clan;
 import pe.com.rc.mobile.model.User;
 import pe.com.rc.mobile.model.clan.BuildClanRequest;
+import pe.com.rc.mobile.model.clan.Clan;
 import pe.com.rc.mobile.service.clan.ClanService;
 import pe.com.rc.mobile.web.util.Constants;
 
@@ -22,6 +22,7 @@ public class ClanController {
 	@Autowired
 	private ClanService clanService;
 
+	// GET MEMBERS BY CLAN
 	@RequestMapping(value = Constants.URL_CLAN
 			+ Constants.URL_GET_CLAN_MEMBERS, method = RequestMethod.POST, produces = { "application/json" })
 	public @ResponseBody List<User> listarLugaresDefault(@RequestBody Clan clan) throws ServiceException {
@@ -38,4 +39,8 @@ public class ClanController {
 	public void buildClan(@RequestBody BuildClanRequest request) throws ServiceException {
 		clanService.buildClan(request);
 	}
+
+	// ACCEPT CLAN MEMBER
+	
+	// DROP CLAN MEMBER
 }
